+++ index.html
@@ -15,6 +15,6 @@
     </form>
     <div class="cardContainer"></div>
     <div class="modalContainer"></div>
-    <script src="script.js"></script>
+    <script type="module" src="./script/index.js"></script>
   </body>
 </html>
