+++ script/apiKey.js
@@ -1,0 +1,9 @@
+export const apiKey = {
+  method: "GET",
+  headers: {
+    accept: "application/json",
+    Authorization:
+      "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIzMmEzYjYwMTg5Mjc3NjliMDFkMjQ0MzExNzlmZTcyNSIsInN1YiI6IjY1OGUxOTQ5ZWEzN2UwMTljZTc0N2EwNCIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.6FU9xe6iFSloZVTTf0OBTX-lRD1jgNk6uvH-hfjeLP4",
+  },
+};
+
